+++ sdk/cosmos/azure_data_cosmos/tests/emulator_tests/mod.rs
@@ -14,6 +14,7 @@ mod cosmos_patch;
 mod cosmos_proxy;
 mod cosmos_query;
 mod cosmos_response_metadata;
+mod cosmos_scope_bug;
 
 #[path = "../framework/mod.rs"]
 mod framework;